+++ CHANGELOG.md
@@ -95,6 +95,7 @@
 - Enhanced `Install-vRSLCMCertificate` cmdlet to check the certificate before attempting to install.
 - Enhanced `Start-vRSLCMUpgrade` cmdlet perform better status checks during the upgrade process.
 - Enhanced `Enable-vRLIContentPack` cmdlet to handle invalid Git Hub tokens, and additional checks to ensure content pack URL is obtained.
+- Enhanced `Deploy-PhotonAppliance` cmdlet to pull the primary datastore and cluster details from SDDC Manager.
 
 ## v2.9.0